requirements.txt
apache-airflow-providers-docker
apache-airflow-providers-http
apache-airflow-providers-apache-spark
psycopg2-binary
confluent-kafka